Add files via upload
So, here we have activated our hacker application. (In reality, this app will be automatically activated and also won't be visible in the frontend). Now, the hacker app has it's affinity to user application (Check the manifest file of user_application wherein task_affinity = "com.activity.user_application"). So now, when the user app is being used, the hacker app will be running in the same task as of use application.